项目进行中

+++ smoc-scheduler-filters/src/main/java/com/smoc/cloud/scheduler/batch/filters/writer/MessageFilterWriter.java
@@ -28,14 +28,14 @@ public class MessageFilterWriter implements ItemWriter<BusinessRouteValue> {
     public void write(List<? extends BusinessRouteValue> list) throws Exception {
 
         Integer count  = list.size();
-//        log.info("[过滤操作]条数：{}:{}", count, System.currentTimeMillis());
-//        Boolean limiter = filtersRedisDataService.limiterMessageFilter("filter:speed:test",10000,10000,1,count);
-//        if(limiter){
+        log.info("[数据加载]条数：{}:{}", count);
+        Boolean limiter = filtersRedisDataService.limiterMessageFilter("filter:speed:test",4000,4000,1,count);
+        if(limiter){
             messageBusinessHandler.handleMessageBusiness(list);
             routeMessageRepository.deleteByBatch(list);
-//        }else {
-//            log.info("[------------触发限流-----------------------]");
-//        }
+        }else {
+            log.info("[------------数据库加载数据----触发限流-----------------------]");
+        }
 
     }
